// src/types.ts
export interface OIDCClientSession {
  /** Session ID */
  sessionId: string;
  /** Session Expires At (Unixtime, ms) */
  sessionExpiresAt: number;

  /** OIDC Code Verifier for PKCE */
  codeVerifier?: string;
  /** OIDC State */
  state?: string;
  /** OIDC Nonce */
  nonce?: string;

  /** OIDC ID Token */
  idToken?: string;
  /** OIDC Access Token */
  accessToken?: string;
  /** OIDC Refresh Token */
  refreshToken?: string;
}

/** Active session data */
export interface OIDCClientActiveSession extends OIDCClientSession {
  idToken: string;
  accessToken: string;
}